worker/python/pyproject.toml
[project]
name = "squid"
version = "0.3.1"
description = "Squid simulation runner python API"
authors = [{ name = "Ben Barber", email = "ben@vivum.ai" }]
requires-python = ">=3.11"
dependencies = ["pyzmq", "orjson"]
classifiers = [
   "Programming Language :: Python :: 3",
   "Operating System :: OS Independent",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
